moved Documentation to xact.templates.impl, because build expects sub path impl
